add package traverser api test
