+++ src/test/java/com/baidu/hugegraph/api/ApiTestSuite.java
@@ -27,6 +27,19 @@ import com.baidu.hugegraph.api.auth.BelongApiTest;
 import com.baidu.hugegraph.api.auth.GroupApiTest;
 import com.baidu.hugegraph.api.auth.TargetApiTest;
 import com.baidu.hugegraph.api.auth.UserApiTest;
+import com.baidu.hugegraph.api.traverser.AllShortestPathsApiTest;
+import com.baidu.hugegraph.api.traverser.CountApiTest;
+import com.baidu.hugegraph.api.traverser.CustomizedPathsApiTest;
+import com.baidu.hugegraph.api.traverser.FusiformSimilarityApiTest;
+import com.baidu.hugegraph.api.traverser.JaccardSimilarityApiTest;
+import com.baidu.hugegraph.api.traverser.NeighborRankApiTest;
+import com.baidu.hugegraph.api.traverser.PersonalRankApiTest;
+import com.baidu.hugegraph.api.traverser.RingsRaysApiTest;
+import com.baidu.hugegraph.api.traverser.SameNeighborsApiTest;
+import com.baidu.hugegraph.api.traverser.ShortestPathApiTest;
+import com.baidu.hugegraph.api.traverser.SingleSourceShortestPathApiTest;
+import com.baidu.hugegraph.api.traverser.TraverserApiTest;
+import com.baidu.hugegraph.api.traverser.WeightedShortestPathApiTest;
 
 @RunWith(Suite.class)
 @Suite.SuiteClasses({
